requirements.txt
# Core image processing
Pillow==10.4.0